Move `NotAuthorizedPage`
