--- client/components/NotAuthorizedPage.stories.tsx
@@ -1,11 +1,0 @@
-import { Story } from '@storybook/react';
-import React from 'react';
-
-import NotAuthorizedPage from './NotAuthorizedPage';
-
-export default {
-	title: 'admin/NotAuthorizedPage',
-	component: NotAuthorizedPage,
-};
-
-export const _default: Story = () => <NotAuthorizedPage />;
