+++ client/views/admin/apps/AppsRoute.tsx
@@ -1,10 +1,10 @@
 import React, { useState, useEffect, FC } from 'react';
 
-import NotAuthorizedPage from '../../../components/NotAuthorizedPage';
 import PageSkeleton from '../../../components/PageSkeleton';
 import { usePermission } from '../../../contexts/AuthorizationContext';
 import { useRouteParameter, useRoute } from '../../../contexts/RouterContext';
 import { useMethod } from '../../../contexts/ServerContext';
+import NotAuthorizedPage from '../../notAuthorized/NotAuthorizedPage';
 import AppDetailsPage from './AppDetailsPage';
 import AppInstallPage from './AppInstallPage';
 import AppLogsPage from './AppLogsPage';
