+++ client/views/admin/import/ImportRoute.js
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import NotAuthorizedPage from '../../../components/NotAuthorizedPage';
 import { usePermission } from '../../../contexts/AuthorizationContext';
+import NotAuthorizedPage from '../../notAuthorized/NotAuthorizedPage';
 import ImportHistoryPage from './ImportHistoryPage';
 import ImportProgressPage from './ImportProgressPage';
 import NewImportPage from './NewImportPage';
