+++ client/views/admin/info/InformationRoute.tsx
@@ -2,13 +2,13 @@ import { Callout, ButtonGroup, Button, Icon } from '@rocket.chat/fuselage';
 import React, { useState, useEffect, memo, ReactElement } from 'react';
 
 import { IStats } from '../../../../definition/IStats';
-import NotAuthorizedPage from '../../../components/NotAuthorizedPage';
 import Page from '../../../components/Page';
 import PageSkeleton from '../../../components/PageSkeleton';
 import { usePermission } from '../../../contexts/AuthorizationContext';
 import { useMethod, useServerInformation, useEndpoint } from '../../../contexts/ServerContext';
 import { useTranslation } from '../../../contexts/TranslationContext';
 import { downloadJsonAs } from '../../../lib/download';
+import NotAuthorizedPage from '../../notAuthorized/NotAuthorizedPage';
 import InformationPage from './InformationPage';
 
 type fetchStatisticsCallback = ((params: { refresh: boolean }) => void) | (() => void);
