+++ client/views/admin/integrations/IntegrationsRoute.js
@@ -1,8 +1,8 @@
 import React, { useMemo } from 'react';
 
-import NotAuthorizedPage from '../../../components/NotAuthorizedPage';
 import { useAtLeastOnePermission } from '../../../contexts/AuthorizationContext';
 import { useRouteParameter } from '../../../contexts/RouterContext';
+import NotAuthorizedPage from '../../notAuthorized/NotAuthorizedPage';
 import IntegrationsPage from './IntegrationsPage';
 import EditIntegrationsPage from './edit/EditIntegrationsPage';
 import OutgoingWebhookHistoryPage from './edit/OutgoingWebhookHistoryPage';
