+++ client/views/admin/invites/InvitesRoute.tsx
@@ -1,7 +1,7 @@
 import React, { ReactElement } from 'react';
 
-import NotAuthorizedPage from '../../../components/NotAuthorizedPage';
 import { usePermission } from '../../../contexts/AuthorizationContext';
+import NotAuthorizedPage from '../../notAuthorized/NotAuthorizedPage';
 import InvitesPage from './InvitesPage';
 
 const InvitesRoute = (): ReactElement => {
