+++ client/views/notAuthorized/NotAuthorizedPage.stories.tsx
@@ -1,0 +1,15 @@
+import { ComponentMeta, ComponentStory } from '@storybook/react';
+import React from 'react';
+
+import NotAuthorizedPage from './NotAuthorizedPage';
+
+export default {
+	title: 'Community/Views/Not Authorized/NotAuthorizedPage',
+	component: NotAuthorizedPage,
+	parameters: {
+		layout: 'fullscreen',
+	},
+} as ComponentMeta<typeof NotAuthorizedPage>;
+
+export const Default: ComponentStory<typeof NotAuthorizedPage> = () => <NotAuthorizedPage />;
+Default.storyName = 'NotAuthorizedPage';
