+++ client/views/omnichannel/directory/OmnichannelDirectoryPage.tsx
@@ -1,7 +1,7 @@
 import { Tabs } from '@rocket.chat/fuselage';
 import React, { useEffect, useCallback, useState, ReactElement } from 'react';
 
-import NotAuthorizedPage from '../../../components/NotAuthorizedPage';
+import NotAuthorizedPage from '../../notAuthorized/NotAuthorizedPage';
 import Page from '../../../components/Page';
 import { usePermission } from '../../../contexts/AuthorizationContext';
 import { useCurrentRoute, useRoute, useRouteParameter } from '../../../contexts/RouterContext';
