+++ client/views/omnichannel/directory/calls/CallTab.tsx
@@ -1,6 +1,6 @@
 import React, { ReactElement } from 'react';
 
-import NotAuthorizedPage from '../../../../components/NotAuthorizedPage';
+import NotAuthorizedPage from '../../../notAuthorized/NotAuthorizedPage';
 import { usePermission } from '../../../../contexts/AuthorizationContext';
 import CallTable from './CallTable';
