+++ ee/client/omnichannel/monitors/MonitorsPageContainer.js
@@ -1,6 +1,6 @@
 import React from 'react';
 
-import NotAuthorizedPage from '../../../../client/components/NotAuthorizedPage';
+import NotAuthorizedPage from '../../../../client/views/notAuthorized/NotAuthorizedPage';
 import PageSkeleton from '../../../../client/components/PageSkeleton';
 import { useHasLicense } from '../../hooks/useHasLicense';
 import MonitorsPage from './MonitorsPage';
